(fix|COS-5179): Improved user flow for adding missing contact email from email participant select view
