+++ packages/apps/spaces/app/organization/[id]/src/components/Tabs/panels/PeoplePanel/ContactCard/ContactCard.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { useParams } from 'next/navigation';
 import { useForm } from 'react-inverted-form';
-import { useRef, useState, MouseEvent } from 'react';
+import { useRef, useState, useEffect, MouseEvent } from 'react';
 
 import { useQueryClient } from '@tanstack/react-query';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
@@ -24,6 +24,7 @@ import { FormAutoresizeTextarea } from '@ui/form/Textarea';
 import { SelectOption } from '@shared/types/SelectOptions';
 import { getGraphQLClient } from '@shared/util/getGraphQLClient';
 import { Card, CardBody, CardHeader } from '@ui/presentation/Card';
+import { useContactCardMeta } from '@organization/src/state/ContactCardMeta.atom';
 import { ConfirmDeleteDialog } from '@ui/overlay/AlertDialog/ConfirmDeleteDialog';
 import { useUpdateContactMutation } from '@organization/src/graphql/updateContact.generated';
 import { useDeleteContactMutation } from '@organization/src/graphql/deleteContact.generated';
@@ -55,14 +56,21 @@ export const ContactCard = ({
   const organizationId = useParams()?.id as string;
   const queryClient = useQueryClient();
   const cardRef = useRef<HTMLDivElement>(null);
-  const [isExpanded, setIsExpanded] = useState(false);
+  const [{ expandedId, initialFocusedField }, setExpandedCardId] =
+    useContactCardMeta();
+  const isExpanded = expandedId === contact.id;
   const [roleIsFocused, setRoleIsFocused] = useState(false);
   const { isOpen, onOpen, onClose } = useDisclosure();
   useOutsideClick({
     ref: cardRef,
-    handler: () => setIsExpanded(false),
+    handler: () => {
+      if (expandedId === contact.id) {
+        setExpandedCardId({ expandedId: undefined, initialFocusedField: null });
+      }
+    },
   });
-
+  const emailInputRef = useRef<HTMLInputElement | null>(null);
+  const nameInputRef = useRef<HTMLInputElement | null>(null);
   const data = ContactFormDto.toForm(contact);
 
   const formId = `contact-form-${data.id}`;
@@ -98,12 +106,31 @@ export const ContactCard = ({
 
   const toggle = (e: MouseEvent<HTMLDivElement>) => {
     if (['name', 'role', 'title'].includes((e.target as HTMLDivElement)?.id)) {
-      setIsExpanded(true);
+      setExpandedCardId({ expandedId: contact.id, initialFocusedField: null });
 
       return;
     }
-    setIsExpanded((prev) => !prev);
+    if (isExpanded) {
+      setExpandedCardId({ expandedId: undefined, initialFocusedField: null });
+    } else {
+      setExpandedCardId({ expandedId: contact.id, initialFocusedField: null });
+    }
   };
+
+  useEffect(() => {
+    if (expandedId === contact.id && initialFocusedField) {
+      if (initialFocusedField === 'name') {
+        nameInputRef.current?.focus();
+
+        return;
+      }
+      if (initialFocusedField === 'email') {
+        emailInputRef.current?.focus();
+
+        return;
+      }
+    }
+  }, [expandedId, initialFocusedField, emailInputRef]);
 
   const prevEmail = data.email;
   const prevPhoneNumberId = data.phoneId;
@@ -278,6 +305,7 @@ export const ContactCard = ({
               h='6'
               name='name'
               formId={formId}
+              ref={nameInputRef}
               placeholder='Name'
               color='gray.700'
               fontWeight='semibold'
@@ -355,6 +383,7 @@ export const ContactCard = ({
               <FormInputGroup
                 formId={formId}
                 name='email'
+                ref={emailInputRef}
                 placeholder='Email'
                 leftElement={<Icons.Mail1 color='gray.500' />}
                 rightElement={
