+++ packages/apps/spaces/app/organization/[id]/src/components/Timeline/events/email/compose-email/EmailMultiCreatableSelect/index.ts
@@ -1,0 +1,1 @@
+export * from './EmailFormMultiCreatableSelect';
