+++ packages/apps/spaces/app/organization/[id]/src/components/Timeline/events/email/compose-email/EmailParticipantSelect.tsx
@@ -5,7 +5,7 @@ import { Flex } from '@ui/layout/Flex';
 import { Text } from '@ui/typography/Text';
 import { emailRegex } from '@organization/src/components/Timeline/events/email/utils';
 
-import { EmailFormMultiCreatableSelect } from './EmailFormMultiCreatableSelect';
+import { EmailFormMultiCreatableSelect } from './EmailMultiCreatableSelect';
 interface EmailParticipantSelect {
   formId: string;
   entryType: string;
